refactor(rust-api): extract shared manifest plumbing into manifest.rs (sc-1639)
Phase 1, slice 1 of the sc-1639 lib.rs Value-cluster split. Moves the shared
manifest read/merge/write/cache helpers out of the 5,323-line lib.rs into a
focused manifest.rs (310 lines): ManifestCache/ManifestCacheKey, the manifest
I/O (mutate/remove/save/load_root/write_atomic/load_entries + write lock),
merge_entries_by_id/merge_object, metadata_modified_ns, strip_jsonc_comments,
and the two manifest consts.

Behavior-preserving move using the sc-1636 pattern: `use super::*` in the new
module, `pub(crate)` on the external surface, and a plain `use manifest::*`
re-export at root so existing lib.rs / jobs.rs / tests.rs call-sites resolve
unchanged. The model-specific resolve_model_manifest_entry /
merge_model_manifest_entry stay at root for the models.rs slice.

lib.rs 5,323 -> 5,017. cargo test (62 rust-api), npm run rust:check, and
cargo build --features embed-web all green.
